src/main/webapp/app/entities/friendrequest/delete/friendrequest-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { FriendrequestService } from '../service/friendrequest.service';

import { FriendrequestDeleteDialogComponent } from './friendrequest-delete-dialog.component';

describe('Friendrequest Management Delete Component', () => {
  let comp: FriendrequestDeleteDialogComponent;
  let fixture: ComponentFixture<FriendrequestDeleteDialogComponent>;
  let service: FriendrequestService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      declarations: [FriendrequestDeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(FriendrequestDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(FriendrequestDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(FriendrequestService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete('ABC');
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith('ABC');
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      })
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
